ソケットの修正

+++ lib/apache/socket_status.go
@@ -6,7 +6,7 @@ type SocketStatus struct {
 	HostID               string `json:"hostid"`
 	VirtualizationSystem string `json:"virtualizationSystem"`
 	// Socket
-	Socket float64 `json:"socket"`
+	Socket int `json:"socket"`
 	// Time
 	Time string `json:"time"`
 	// Error
